+++ src/Ajedrez/Main.cpp
@@ -5,7 +5,7 @@
 // #undef main
 
 const int FPS = 60;
-const int DELAY_TIME = 1000.0f / FPS;
+const int DELAY_TIME = int(1000.0f / FPS);
 
 int main(int argc, char* args[])
 {
