+++ src/Ajedrez/PauseState.cpp
@@ -44,7 +44,7 @@ bool PauseState::OnExit()
 
 	m_gameObjects.clear();
 	// clear the texture manager
-	for (int i = 0; i < m_textureIDList.size(); i++)
+	for (size_t i = 0; i < m_textureIDList.size(); i++)
 	{
 		TextureManager::Instance()->ClearFromTextureMap(m_textureIDList[i]);
 	}
@@ -71,7 +71,7 @@ void PauseState::SetCallbacks(const std::vector<Callback>& callbacks)
 	// go through the game objects
 	if (!m_gameObjects.empty())
 	{
-		for (int i = 0; i < m_gameObjects.size(); i++)
+		for (size_t i = 0; i < m_gameObjects.size(); i++)
 		{
 			// if they are of type MenuButton then assign a callback based on the id passed in from the file
 			MenuButton* pButton = dynamic_cast<MenuButton*>(m_gameObjects[i]);
